Разбивает проект на модули.
