+++ src/components/days.js
@@ -1,0 +1,13 @@
+import {getDayComponent} from './day.js';
+
+const getDaysComponent = () => {
+  return `
+    <ul class="trip-days">
+      ${getDayComponent()}
+    </ul>
+  `;
+};
+
+export {
+  getDaysComponent
+};
