+++ src/components/events-list.js
@@ -1,0 +1,15 @@
+import {getEventsComponent} from './events.js';
+
+const EVENTS_NUMBER = 3;
+
+const getEventsListComponent = () => {
+  return `
+    <ul class="trip-events__list">
+      ${getEventsComponent(EVENTS_NUMBER)}
+    </ul>
+  `;
+};
+
+export {
+  getEventsListComponent
+};
