+++ src/components/events.js
@@ -1,0 +1,11 @@
+import {getEventComponent} from './event.js';
+
+const getEventsComponent = (eventsNumber) => {
+  return new Array(eventsNumber)
+    .fill(getEventComponent())
+    .join(``);
+};
+
+export {
+  getEventsComponent
+};
