+++ src/utils.js
@@ -1,0 +1,7 @@
+const renderComponent = (container, component, position = `beforeend`) => {
+  container.insertAdjacentHTML(position, component);
+};
+
+export {
+  renderComponent
+};
